rename to attackers and saliency_interpreters
