+++ allennlp/interpret/attackers/__init__.py
@@ -1,3 +1,3 @@
-from allennlp.interpret.attack.attacker import Attacker
-from allennlp.interpret.attack.input_reduction import InputReduction
-from allennlp.interpret.attack.hotflip import Hotflip
+from allennlp.interpret.attackers.attacker import Attacker
+from allennlp.interpret.attackers.input_reduction import InputReduction
+from allennlp.interpret.attackers.hotflip import Hotflip
