+++ allennlp/interpret/saliency_interpreters/__init__.py
@@ -1,4 +1,4 @@
-from allennlp.interpret.saliency.saliency_interpreter import SaliencyInterpreter
-from allennlp.interpret.saliency.simple_gradient import SimpleGradient
-from allennlp.interpret.saliency.integrated_gradient import IntegratedGradient
-from allennlp.interpret.saliency.smooth_gradient import SmoothGradient
+from allennlp.interpret.saliency_interpreters.saliency_interpreter import SaliencyInterpreter
+from allennlp.interpret.saliency_interpreters.simple_gradient import SimpleGradient
+from allennlp.interpret.saliency_interpreters.integrated_gradient import IntegratedGradient
+from allennlp.interpret.saliency_interpreters.smooth_gradient import SmoothGradient
