+++ allennlp/interpret/saliency_interpreters/integrated_gradient.py
@@ -2,7 +2,7 @@ import math
 from typing import List, Dict 
 import numpy
 from allennlp.common.util import JsonDict, sanitize
-from allennlp.interpret.saliency import SaliencyInterpreter
+from allennlp.interpret.saliency_interpreters import SaliencyInterpreter
 from allennlp.predictors import Predictor
 from allennlp.modules.text_field_embedders import TextFieldEmbedder
 from allennlp.data import Instance
